Create Stack.c
A stack is a linear data structure that follows the principle of Last In First Out (LIFO). This means the last element inserted inside the stack is removed first. You can think of the stack data structure as the pile of plates on top of another. And, if you want the plate at the bottom, you must first remove all the plates on top.